requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
WTForms==3.1.0
Werkzeug==3.0.1
SQLAlchemy==2.0.23
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
Pillow==10.1.0
email-validator==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
blinker==1.7.0
alembic==1.13.1
Mako==1.3.0